requirements.txt
python-dotenv
azure-identity
python-docx
semantic_kernel
